created app.py

+++ src/pipeline/predict_pipline.py
@@ -20,11 +20,11 @@ class Predcict_Pipeline:
             return preds
         except Exception as e:
             raise CustomException(e, sys)
-        
+
 
 class CustomData:
     def __init__(self, gender: str, race_ethnicity: str, parental_level_of_education, lunch: str,
-                 test_preparation_course: str, reading_score: int, writing_score: int):
+                 test_preparation_course: str, reading_score: float, writing_score: float):
         self.gender = gender
         self.race_ethnicity = race_ethnicity
         self.parental_level_of_education = parental_level_of_education
